+++ CHANGELOG.md
@@ -5,6 +5,8 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [Unreleased]
+
+## [3.0.0] - 2016-10-27
 ### Added
 - Custom logger decorators can be registered on the Factory and applied via config
 - New `Decorator\DefaultContext` which allows default context values to be set for all logs.